% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple-normal.R
\name{one.sim.diff.normal}
\alias{one.sim.diff.normal}
\title{Generate fake data for one subject, difference by condition of normally distributed outcome variable}
\usage{
one.sim.diff.normal(trash = NULL, mean1, mean2, sd1, sd2 = sd1, n.trials1,
  n.trials2 = n.trials1)
}
\arguments{
\item{trash}{A throwaway first argument to allow piping from apply functions like replicate.}

\item{mean1}{A numeric value specifying the "true mean" of the first
task condition from which simulated data will be generated.}

\item{mean2}{A numeric value specifying the "true mean" of the second
task condition from which simulated data will be generated.}

\item{sd1}{A numeric value specifying the "true standard deviation"
of the first task condition from which simulated data will be generated.}

\item{sd2}{A numeric value specifying the "true standard deviation"
of the second task condition from which simulated data will be generated.
Defaults to \code{sd1}, aka assumes equal variance between conditions.}

\item{n.trials1}{A numeric value specifying the number of trials to
be presented in the first task condition.}

\item{n.trials2}{A numeric value specifying the number of trials to
be presented in the second task condition. Defaults to \code{n.trials1},
aka assumes equal number of trials in both conditions.}
}
\value{
The estimated difference and standard error of the difference
estimate from one round of simulated data.
}
\description{
Calculates the difference of a normally distributed outcome variable by condition for a single subject.
}
