% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple-normal.R
\name{one.sim.mean.normal}
\alias{one.sim.mean.normal}
\title{Generate fake data for one subject, one normally distributed outcome variable}
\usage{
one.sim.mean.normal(trash = NULL, mean = 0, sd = 1, n.trials)
}
\arguments{
\item{trash}{A throwaway first argument to allow piping from apply functions like replicate.}

\item{mean}{A numeric value specifying the "true mean" from which
simulated data will be generated.}

\item{sd}{A numeric value specifying the "true standard deviation"
from which simulated data will be generated.}

\item{n.trials}{A numeric value specifying the number of trials to
be presented.}
}
\value{
The estimated mean and standard error of the mean estimate
from one round of simulated data.
}
\description{
Calculates the mean of one normally distributed outcome variable for a single subject.
}
