% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psychophysics-2afc.R
\name{one.sim.psyphys2afc}
\alias{one.sim.psyphys2afc}
\title{Generate fake data for one subject, 2AFC psychophysics-style task}
\usage{
one.sim.psyphys2afc(trash = NULL, xmin = NULL, xmax = NULL,
  xstep = NULL, n.xlevels = NULL, mean, sd, trials.per.bin)
}
\arguments{
\item{trash}{A throwaway first argument to allow piping from apply functions like replicate.}

\item{mean}{A numeric value specifying the "true mean" from which
simulated data will be generated.}

\item{sd}{A numeric value specifying the "true standard deviation"
from which simulated data will be generated.}

\item{trials.per.bin}{A numeric value specifying the number of trials
presented PER stimulus intensity bin.}

\item{bins}{A vector of numeric values, indicating the levels
of intensity at which stimuli will be presented.}
}
\value{
The estimated mean and standard deviation of the probit model
fit to one round of simulated data.
}
\description{
Fits a probit/cumulative normal model to fake single-subject data.
}
