% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdt-dprime.R
\name{one.sim.sdt.dprime}
\alias{one.sim.sdt.dprime}
\title{Generate fake data for one subject, SDT task, calculating d'}
\usage{
one.sim.sdt.dprime(trash = NULL, dprime, sdt.c = 0, n.trials.signal,
  n.trials.noise)
}
\arguments{
\item{trash}{A throwaway first argument to allow piping from apply functions like replicate.}

\item{dprime}{A numeric value specifying the "true d'" from which
simulated data will be generated.}

\item{sdt.c}{A numeric value specifying the "true c" (criterion parameter)
from which simulated data will be generated. Defaults to 0 (no bias).}

\item{n.trials.signal}{A numeric value specifying the number of trials
to be presented that contain signal (correct response is "yes").}

\item{n.trials.noise}{A numeric value specifying the number of trials
to be presented that contain noise (correct response is "no").}
}
\value{
The estimated d' and c from one round of simulated data. Calculated
on Snodgrass-corrected simulated hit and FA rates (Snodgrass & Corwin, 1988).
}
\description{
Calculates d' from a fake SDT task for a single subject.
}
