% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdt-dprime.R
\name{snodgrass}
\alias{snodgrass}
\title{internal
Calculates rate correction against ceiling/floor rates
per Snodgrass & Corwin, 1988.}
\usage{
snodgrass(num, denom)
}
\arguments{
\item{num}{numerator}

\item{denom}{denominator}
}
\description{
internal
Calculates rate correction against ceiling/floor rates
per Snodgrass & Corwin, 1988.
}
