% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-replicate.R
\name{replicate.sim.by.param}
\alias{replicate.sim.by.param}
\title{Repeat data simulation for multiple levels of a parameter}
\usage{
replicate.sim.by.param(type = "psychophysics.2afc", ...)
}
\arguments{
\item{type}{A string indicating the label for the type of
task to be replicated. Currently supports \code{"psychophysics.2afc"}.}

\item{...}{arguments ultimately passed to the workhorse single simulation
function specified by \code{type}.
For \code{"psychophysics.2afc"}, please see the help for \code{one.sim.psyphys2afc}.}
}
\value{
a long dataframe containing the performance parameter estimates
from each simulated task run, labeled by the varying parameter of interest.
}
\description{
Repeat data simulation for multiple levels of a parameter
}
