% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-replicate.R
\name{replicate.sim}
\alias{replicate.sim}
\title{internal
workhorse function to repeatedly call a \code{one.sim} function
and store the results of every call}
\usage{
replicate.sim(n.reps, type = "psychophysics.2afc", ...)
}
\arguments{
\item{n.reps}{number of simulations to run}

\item{type}{A string indicating the label for the type of
task to be replicated. Currently supports \code{"psychophysics.2afc"}.}

\item{...}{arguments ultimately passed to the workhorse single simulation
function specified by \code{type}.
For \code{"psychophysics.2afc"}, please see the help for \code{one.sim.psyphys2afc}.}
}
\value{
a dataframe where each row is a simulation and each column
is an estimated parameter.
}
\description{
internal
workhorse function to repeatedly call a \code{one.sim} function
and store the results of every call
}
